+++ src/RLArduinoInstrument.h
@@ -18,7 +18,7 @@ Revision History
 #include "RLArduinoSerial.h"
 #include <SPI.h>
 
-#define RL_ARDUINO_INSTRUMENT_LIB_VERSION  (F("1.4.0"))
+#define RL_ARDUINO_INSTRUMENT_LIB_VERSION  (F("1.7.0"))
 
 //Error definitions
 #define ERROR_NONE 0
@@ -33,7 +33,7 @@ typedef uint16_t command_t;
 typedef uint16_t index_t;
 
 //global variable
-const char c_noData = '\0';
+const char c_noData = '\x04';
 
 //Command structure
 struct SerialCommand {
@@ -49,6 +49,12 @@ class RLArduinoInstrument:public RLArduinoSerial {
     void addCommand(command_t command, const char* description, error_t (*commandHandler)(index_t index));
     error_t executeCommand(command_t command);
     bool getDebug() {return g_debug;}
+    template <typename T> 
+    void print(T value);
+    void print(float value, int precision = 3);
+    template <typename T> 
+    void println(T value);
+    void println(float value, int precision = 3);
     void printCommands();
     void printCommandDescription(index_t index);
     void printError(error_t error);
@@ -57,14 +63,16 @@ class RLArduinoInstrument:public RLArduinoSerial {
     template <typename T> 
     void printLabelValue(error_t error, String label,  T value);
     void printLabelValue(error_t error, String label,  float value, int precision = 3);
-
     template <typename T> 
     void printResult(error_t error, index_t index,  T value);
     void printResult(error_t error, index_t index,  float value, int precision = 3);
     void printResult(index_t index);
+    error_t requestFloat(float &value, String prompt,int precision = 3);
     error_t requestFloat(float &value, index_t index,int precision = 3);
     error_t requestLong(long &value, index_t index);
+    error_t requestLong(long &value, String prompt);
     error_t requestString(String &value, index_t index);
+    error_t requestString(String &value,String prompt);
     void setDebug(bool debug) {g_debug = debug;}
   private:
     //global variables