+++ CHANGELOG.md
@@ -22,3 +22,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [1.4.0] - 2024-04-05
 - added precision to the requestFloat function
+
+## [1.5.0] - 2024-04-14
+- added new versions of requestLong, requestFloat and requestString that pass a prompt string
+
+## [1.6.0] - 2024-05-03
+- added the print and println functions
+
+## [1.7.0] - 2024-08-21
+- modified the executeCommand function to return a c_data character if the command is not found